add email method, tested app, email
