+++ index.js
@@ -4,4 +4,14 @@ function Deta(projectKey, host) {
   return new _Deta(projectKey, host);
 }
 
-module.exports = Deta;
+let exportObj = { Deta };
+
+try {
+  const { App } = require("detalib"); // how to import from a layer?
+  exportObj.app = App();
+  exportObj.App = App;
+} catch (e) {
+
+}
+
+module.exports = exportObj;
